Move desktop software API to legacy subpackage
