+++ backend/apps/core/routes/root.py
@@ -32,7 +32,7 @@ urlpatterns = [
     path('api/v1/', include('apps.company.routes.api')),
     path('api/v1/', include('apps.xl_dashboard.routes.api')),
     path('api/v1/', include('apps.software.routes.api')),
-    path('api/v1/', include('apps.software.desktop_software_urls')),
+    path('api/v1/', include('apps.software.legacy.desktop_software_urls')),
     path('api/v1/', include('apps.filehost.routes.api')),
     path('api/v1/xlmine/', include('apps.xlmine.routes.api')),
     path('api/v1/confirmation-code/', include('apps.confirmation.routes.api')),
