+++ backend/apps/software/legacy/desktop_software_urls.py
@@ -1,7 +1,11 @@
 # software/desktop_software_urls.py
 from django.urls import path
 
-from apps.software.desktop_software_api import software_auth, set_user_hw_id, get_software_version
+from apps.software.legacy.desktop_software_api import (
+    software_auth,
+    set_user_hw_id,
+    get_software_version,
+)
 
 urlpatterns = [
     path('login_program/', software_auth, name='software_auth'),
